+++ CHANGELOG.md
@@ -9,6 +9,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
+## [0.4.0] - 2016-08-21
+### Changed
+- DF-681 Event firing changes for resources.
+- DF-607 Making service docs always viewable, even auto-generated ones.
+### Fixed
+- Wrapper not affecting stored procedure output when no output parameters are present.
+
 ## [0.3.2] - 2016-07-08
 ### Added
 - DF-636 Adding ability using 'ids' parameter to return the schema of a stored procedure or function
@@ -88,7 +95,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## 0.1.0 - 2015-10-24
 First official release working with the new [df-core](https://github.com/dreamfactorysoftware/df-core) library.
 
-[Unreleased]: https://github.com/dreamfactorysoftware/df-sqldb/compare/0.3.2...HEAD
+[Unreleased]: https://github.com/dreamfactorysoftware/df-sqldb/compare/0.4.0...HEAD
+[0.3.2]: https://github.com/dreamfactorysoftware/df-sqldb/compare/0.3.2...0.4.0
 [0.3.2]: https://github.com/dreamfactorysoftware/df-sqldb/compare/0.3.1...0.3.2
 [0.3.1]: https://github.com/dreamfactorysoftware/df-sqldb/compare/0.3.0...0.3.1
 [0.3.0]: https://github.com/dreamfactorysoftware/df-sqldb/compare/0.2.2...0.3.0